requirements.txt
Django
djangorestframework
django-cors-headers
Pillow
djangorestframework-simplejwt
djangorestframework django-filter